Reduce enemy health every second while being shot by the laser beamer.

+++ Tower Defense Brackeys/Assets/Scripts/Enemy.cs
@@ -3,8 +3,7 @@
 public class Enemy : MonoBehaviour 
 {
     public float speed = 10f;
-
-    public int health = 100;
+    public float health = 100f;
 
     public int moneyReward = 50;
 
@@ -18,7 +17,7 @@ public class Enemy : MonoBehaviour
         target = Waypoints.points[0];
     }
 
-    public void TakeDamage(int amount)
+    public void TakeDamage(float amount)
     {
         health -= amount;
